refactor: extract PageGamer which runs the game
